// types/ICompanyViewItem.ts
// generated typescript definitions from database using groovy script

export interface ICompanyViewItem {
  company_number: string
  name: string
  streetaddress: string
  county: string
  area: string
  country: string
  postcode: string
  company_type: string
  country_of_origin: string
  status: string
  date_of_creation: number
  updated: Date
  can_file: string
  balance_sheet_date: Date
  accountants: string
  employees: string
  current_assets: string
  cash_at_bank: string
  debtors: string
  creditors: string
  fixed_assets: string
  net_assets: string
  total_assets_less_current_liabilities: string
  equity: string
  revenue: string
  profit: string
  officers: string[]
}